# Make execution context struct and other performance tweaks