--- a/Jint/Native/Error/ErrorInstance.cs
+++ b/Jint/Native/Error/ErrorInstance.cs
@@ -6,17 +6,9 @@ namespace Jint.Native.Error
     public class ErrorInstance : ObjectInstance
     {
         public ErrorInstance(Engine engine, string name)
-            : base(engine)
+            : base(engine, objectClass: "Error")
         {
             FastAddProperty("name", name, true, false, true);
-        }
-
-        public override string Class
-        {
-            get
-            {
-                return "Error";
-            }
         }
 
         public override string ToString()