This is a kind of a collection of different tweaks as I was going through performance profiling, so sorry for the big PR. Main idea is to reduce allocations with ExecutionContext being a struct and then various tweaks like accessing fieds instead of properties when possible. Benchmark also showed that using type field check for null and undefined is faster than ReferenceEquals calls.

* ensure method bodies are inline friendly
* remove virtual calls from AsNumber, AsString and AsSymbol
  * direct field access when possible
  * offer migration path via extension methods
* use IsUndefined & IsNull which are faster than ReferenceEquals


## ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Slice|100|436.2 us|161.1328|660.16 KB|
| **New** |	|	| **432.7 us (-1%)** | **161.1328 (0%)** | **660.16 KB (0%)** |
| Old |Concat|100|468.4 us|175.7813|720.31 KB|
| **New** |	|	| **454.0 us (-3%)** | **175.7813 (0%)** | **720.31 KB (0%)** |
| Old |Unshift|100|17,912.2 us|3562.5000|14672.66 KB|
| **New** |	|	| **18,052.9 us (+1%)** | **3562.5000 (0%)** | **14672.66 KB (0%)** |
| Old |Push|100|10,861.8 us|343.7500|1438.28 KB|
| **New** |	|	| **10,491.0 us (-3%)** | **343.7500 (0%)** | **1438.28 KB (0%)** |
| Old |Index|100|12,106.5 us|390.6250|1637.5 KB|
| **New** |	|	| **11,950.4 us (-1%)** | **390.6250 (0%)** | **1637.5 KB (0%)** |
| Old |Map|100|3,382.7 us|765.6250|3149.22 KB|
| **New** |	|	| **3,143.6 us (-7%)** | **691.4063 (-10%)** | **2833.59 KB (-10%)** |
| Old |Apply|100|569.2 us|188.4766|774.22 KB|
| **New** |	|	| **564.2 us (-1%)** | **188.4766 (0%)** | **774.22 KB (0%)** |
| Old |JsonStringifyParse|100|4,523.7 us|1273.4375|5225 KB|
| **New** |	|	| **4,489.3 us (-1%)** | **1273.4375 (0%)** | **5225.78 KB (0%)** |


## ArrayStressBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|608.7 ms|60000.0000|2937.5000|245.9 MB|
| **New** |	|	| **585.3 ms (-4%)** | **60000.0000 (0%)** | **2937.5000 (0%)** | **245.9 MB (0%)** |


## DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|78.48 ms|1500.0000|500.0000|-|7747.67 KB|
| **New** |	|	| **75.73 ms (-4%)** | **1375.0000 (-8%)** | **250.0000 (-50%)** | **-** | **7662.21 KB (-1%)** |
| Old |Run|dromaeo-core-eval|18.03 ms|62.5000|-|-|299.02 KB|
| **New** |	|	| **16.32 ms (-9%)** | **62.5000 (0%)** | **-** | **-** | **298.7 KB (0%)** |
| Old |Run|dromaeo-object-array|163.17 ms|42562.5000|2000.0000|1000.0000|178312.87 KB|
| **New** |	|	| **158.04 ms (-3%)** | **42437.5000 (0%)** | **1937.5000 (-3%)** | **937.5000 (-6%)** | **178314.09 KB (0%)** |
| Old |Run|dromaeo-object-regexp|892.92 ms|52812.5000|31375.0000|21250.0000|413157.83 KB|
| **New** |	|	| **882.55 ms (-1%)** | **54687.5000 (+4%)** | **33500.0000 (+7%)** | **22750.0000 (+7%)** | **417652.55 KB (+1%)** |
| Old |Run|dromaeo-object-string|1,214.16 ms|219375.0000|176187.5000|173375.0000|1393091.9 KB|
| **New** |	|	| **1,194.33 ms (-2%)** | **219437.5000 (0%)** | **176125.0000 (0%)** | **173625.0000 (0%)** | **1394577.96 KB (0%)** |
| Old |Run|dromaeo-string-base64|179.78 ms|6125.0000|312.5000|-|26272.45 KB|
| **New** |	|	| **175.07 ms (-3%)** | **6375.0000 (+4%)** | **312.5000 (0%)** | **-** | **27381.53 KB (+4%)** |


## EvaluationBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Jint|20|710.5 us|120.1172|492.5 KB|
| **New** |	|	| **686.9 us (-3%)** | **119.1406 (-1%)** | **488.91 KB (-1%)** |


## LinqJsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Jint|10|69.97 ms|5062.5000|2500.0000|-|30.56 MB|
| **New** |	|	| **71.57 ms (+2%)** | **5125.0000 (+1%)** | **2500.0000 (0%)** | **125.0000** | **30.57 MB (0%)** |


## MinimalScriptBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Jint|10|19.10 us|8.1482|33.44 KB|
| **New** |	|	| **19.47 us (+2%)** | **8.1482 (0%)** | **33.44 KB (0%)** |


## StopwatchBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|1|1.213 s|17125.0000|62.5000|68.5 MB|
| **New** |	|	| **1.169 s (-4%)** | **14375.0000 (-16%)** | **62.5000 (0%)** | **57.73 MB (-16%)** |


## SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|620.7 ms|12000.0000|312.5000|-|50.05 MB|
| **New** |	|	| **601.3 ms (-3%)** | **11875.0000 (-1%)** | **312.5000 (0%)** | **-** | **49.56 MB (-1%)** |
| Old |Run|3d-morph|588.6 ms|11812.5000|2500.0000|1062.5000|68.03 MB|
| **New** |	|	| **581.5 ms (-1%)** | **11812.5000 (0%)** | **2500.0000 (0%)** | **1062.5000 (0%)** | **68.03 MB (0%)** |
| Old |Run|3d-raytrace|503.0 ms|22500.0000|625.0000|62.5000|92.11 MB|
| **New** |	|	| **485.4 ms (-3%)** | **21875.0000 (-3%)** | **750.0000 (+20%)** | **125.0000 (+100%)** | **89.96 MB (-2%)** |
| Old |Run|access-binary-trees|207.1 ms|17687.5000|375.0000|62.5000|73 MB|
| **New** |	|	| **204.7 ms (-1%)** | **16500.0000 (-7%)** | **437.5000 (+17%)** | **62.5000 (0%)** | **68.51 MB (-6%)** |
| Old |Run|access-fannkuch|1,916.6 ms|9437.5000|187.5000|-|37.87 MB|
| **New** |	|	| **1,880.4 ms (-2%)** | **9437.5000 (0%)** | **125.0000 (-33%)** | **-** | **37.87 MB (0%)** |
| Old |Run|access-nbody|598.4 ms|13125.0000|-|-|52.6 MB|
| **New** |	|	| **556.6 ms (-7%)** | **13062.5000 (0%)** | **-** | **-** | **52.42 MB (0%)** |
| Old |Run|access-nsieve|738.8 ms|15375.0000|3937.5000|1875.0000|73.09 MB|
| **New** |	|	| **704.6 ms (-5%)** | **15375.0000 (0%)** | **3937.5000 (0%)** | **1875.0000 (0%)** | **73.09 MB (0%)** |
| Old |Run|bitops-3bit-bits-in-byte|437.9 ms|19750.0000|-|-|79.06 MB|
| **New** |	|	| **418.3 ms (-4%)** | **18500.0000 (-6%)** | **-** | **-** | **74.18 MB (-6%)** |
| Old |Run|bitops-bits-in-byte|730.5 ms|12437.5000|62.5000|-|49.86 MB|
| **New** |	|	| **690.4 ms (-5%)** | **11562.5000 (-7%)** | **62.5000 (0%)** | **-** | **46.44 MB (-7%)** |
| Old |Run|bitops-bitwise-and|438.7 ms|13562.5000|-|-|54.32 MB|
| **New** |	|	| **422.5 ms (-4%)** | **13562.5000 (0%)** | **-** | **-** | **54.32 MB (0%)** |
| Old |Run|bitops-nsieve-bits|761.7 ms|21625.0000|250.0000|62.5000|88.65 MB|
| **New** |	|	| **724.2 ms (-5%)** | **21625.0000 (0%)** | **250.0000 (0%)** | **62.5000 (0%)** | **88.65 MB (0%)** |
| Old |Run|controlflow-recursive|297.1 ms|25937.5000|62.5000|-|103.78 MB|
| **New** |	|	| **293.4 ms (-1%)** | **23625.0000 (-9%)** | **-** | **-** | **94.61 MB (-9%)** |
| Old |Run|crypto-aes|461.1 ms|5625.0000|250.0000|-|24.51 MB|
| **New** |	|	| **445.5 ms (-3%)** | **5500.0000 (-2%)** | **312.5000 (+25%)** | **62.5000** | **24.16 MB (-1%)** |
| Old |Run|crypto-md5|340.7 ms|30562.5000|187.5000|-|124.99 MB|
| **New** |	|	| **330.7 ms (-3%)** | **29500.0000 (-3%)** | **312.5000 (+67%)** | **62.5000** | **120.95 MB (-3%)** |
| Old |Run|crypto-sha1|328.2 ms|24875.0000|312.5000|-|101.48 MB|
| **New** |	|	| **322.6 ms (-2%)** | **23875.0000 (-4%)** | **187.5000 (-40%)** | **-** | **97.36 MB (-4%)** |
| Old |Run|date-format-tofte|353.3 ms|12625.0000|-|-|50.73 MB|
| **New** |	|	| **327.6 ms (-7%)** | **12500.0000 (-1%)** | **62.5000** | **-** | **50.32 MB (-1%)** |
| Old |Run|date-format-xparb|336.1 ms|10375.0000|250.0000|-|42.27 MB|
| **New** |	|	| **329.4 ms (-2%)** | **10062.5000 (-3%)** | **250.0000 (0%)** | **-** | **41.1 MB (-3%)** |
| Old |Run|math-cordic|965.9 ms|22562.5000|62.5000|-|90.25 MB|
| **New** |	|	| **912.8 ms (-5%)** | **21312.5000 (-6%)** | **62.5000 (0%)** | **-** | **85.48 MB (-5%)** |
| Old |Run|math-partial-sums|304.1 ms|7250.0000|-|-|29.04 MB|
| **New** |	|	| **292.2 ms (-4%)** | **7250.0000 (0%)** | **-** | **-** | **29.03 MB (0%)** |
| Old |Run|math-spectral-norm|363.0 ms|18062.5000|-|-|72.31 MB|
| **New** |	|	| **344.4 ms (-5%)** | **16875.0000 (-7%)** | **-** | **-** | **67.63 MB (-6%)** |
| Old |Run|regexp-dna|342.1 ms|2312.5000|1875.0000|1562.5000|13.55 MB|
| **New** |	|	| **341.2 ms (0%)** | **2187.5000 (-5%)** | **1875.0000 (0%)** | **1500.0000 (-4%)** | **13.55 MB (0%)** |
| Old |Run|string-base64|271.9 ms|8875.0000|500.0000|-|37.16 MB|
| **New** |	|	| **265.3 ms (-2%)** | **9312.5000 (+5%)** | **375.0000 (-25%)** | **-** | **38.79 MB (+4%)** |
| Old |Run|string-fasta|452.5 ms|27125.0000|62.5000|-|108.53 MB|
| **New** |	|	| **426.7 ms (-6%)** | **26750.0000 (-1%)** | **-** | **-** | **107.01 MB (-1%)** |
| Old |Run|string-tagcloud|212.6 ms|14562.5000|1750.0000|625.0000|67.2 MB|
| **New** |	|	| **206.3 ms (-3%)** | **14375.0000 (-1%)** | **1875.0000 (+7%)** | **750.0000 (+20%)** | **65.99 MB (-2%)** |
| Old |Run|string-unpack-code|163.3 ms|14062.5000|3937.5000|937.5000|68.64 MB|
| **New** |	|	| **162.1 ms (-1%)** | **13687.5000 (-3%)** | **4500.0000 (+14%)** | **937.5000 (0%)** | **64.88 MB (-5%)** |
| Old |Run|string-validate-input|184.7 ms|7375.0000|312.5000|62.5000|34.26 MB|
| **New** |	|	| **179.4 ms (-3%)** | **7437.5000 (+1%)** | **500.0000 (+60%)** | **125.0000 (+100%)** | **34.08 MB (-1%)** |


## UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|349.9 ms|52312.5000|875.0000|210.48 MB|
| **New** |	|	| **332.4 ms (-5%)** | **49125.0000 (-6%)** | **125.0000 (-86%)** | **196.69 MB (-7%)** |